backend/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langgraph
sqlalchemy
pydantic
pydantic-settings
python-dotenv
